Created a cleaner orbitControl property
I wanted to access the enable property for orbitControls, but orbitControls is the value of a another property within the scene object. To make this cleaner, an 'attach' property was added to the orbitControls component (r3f) to spawn a more visible property in the scene object as can be witnessed in the console (developer tools).