Dragging model as a whole (w/o dismembering it)
I made the linking members (subcomponents) of the model (both) inextricable from each other so they can be dragged as a single unit. This was done with the transformGroup property which can be found in three docs, a per usual. Unfortunately, it only works for a single component at a time, so Dragable had to be used on the models separately as seen in App.js.