Added the ground work for the CameraControls comp
CameraControls is for controlling the angles at which you the user views the scene through the lens of the camera as the name implies.
Along with the camera component comes the state file which I will get to later.
To understand how you're going to accomplish the repositioning of the camera, with animation, you need a recap and/or principle understanding of a few concepts.
Vectors
Remember vectors are different from coordinates. Vector values tell you how to move from one end of the vector (tail) to the other end (tip). This can be in a 2 dimensional or 3-dimensional space. You are working with 3 of them at this point in the project: the vector of the camera, the vector of the angle you want looking onto the first car (right side), and the angle/ vector looking onto the second car (LS). The last 2 of the aforementioned 3 vectors will be managed in the state file.
We then extract the camera  - via destructuring - and chain its position property which is of interest. The position so happens to be a vector3 so we can chain lerp - a function that can be found in threejs docs, a usual - that will give us our animation effect moving from one vector to another. Lerp takes two arguments: the value of the target vector you want to interpolate towards (move) and the speed. At this point, the camera will snap to the designated camera position anytime you orbit because every frame the instructions in the useFrame is executed.