Adding logic to the camera movements
You want the camera to lock in on a certain vector direction conditionally. So you created a condition in which the camera locks in place - smooth animation/ transition included - only when true. In the end, a button should be able to control when to engage in  behavior befitting the truthy expression.

Note the usage of Lerp again on the vector3 property 'target' because we want a smooth transition again but to the point about which the camera rotates

Note also, you must update after every change of the camera's transform, as stated in the documentation

Didn't want to mutate the position property, a vector3 object if you're thinking in terms of standard javascript "literal" objects, so you "cloned" the position vector.

Added an arrowHelper for reference of the camera direction vector.

You can firmly say that you learned Method chaining and gained a good understanding of "this". Visit notion for a short crash course but in essence, they're responsible for the behavior where methods properties don't need to immediately trail the objects from which the hail(come from), in the dot notation.

At the start, we talked about constraining the view. We will want to leave that constrained view at some point, so we wrote condition for that too.