User interface for CameraControls
The gist of what this component does is to give the user command of the state of the camera. It's simply a button (two actually).
Two objects in pairs of vectors are provided. One pair for the left car and one for the right. The objects are injected into a handler respective of button sides it should feed.