Car colors can be changed. No way.
You updated colorSelector’s handleClick function to work with the cars’ “objects” and not the square “object” from an earlier time. The cars’ objects will so happen live in the activeMesh property of the state “object.”

Where did the name property come from in the array of objects in CameraButtons? You needed a way to “access” each car “object”. As far as I can remember, “accessing” components is not possible in react. The trick for getting the name property of the cars is to disable transformGroup, “log” the car “object” so it appears in console when dragged, drag each cars’ shell noting the names’ association to a specific car, placing names in respective “objects” of sets array and restoring truthy to transformGroup. Tah Dah.

You wanted to set the activeMesh property to whatever car the arrow buttons handle because only then would ColorSelector have a value to work with from activeMesh. This means calling a method on Object3d or related that returns an object whose name matches that provided in the method, think querySelector from javascript. Unfortunately, this code has to be split between two components as the CameraButton component doesn’t sit in canvas thus the “scene” object won’t work. You’ll see that this comes with another upside in that your code in CameraControl, the partnering component, has already extracted the “scene” object via useFrame.