Applying shadows that look the part
The imported model has many subcomponents so we want to check them for a mesh and apply shadow. How do we check them all, we use a traverse method that iterates - similar forEach or map but not sure of the order - over all the properties of the object it is called on. In fact, traverse even iterates over the child of the child. Don't ask but useLoader returns an object with scene as it’s property which works🙂. All that was left to do, adding the shadows. The higher the value of the shadow properties the sharper/ smoother the shadow👍
Note. The car had a weird meshy shadow on its side, THREE. FrontSide counteracts that.