Droid in motion
The aim is to import a model and access its animations using the Animation system of threejs.

Why not create the animations for the droid from scratch in threejs?

It is a lot easier to download or export a model that has animations. I mean look at all the details of the robot - the little arms and flaps - could you animate them in threejs?

The animation system is to a sound engineer’s (or intuitively a dj) mixer console it even has a play button similar to the actual thing.

The impetus for the design/architecture of your Animation system code but through an analogy first:

An order is placed for a car. That’s the component. Its function is like the car going through all the processes to create the final product. All the cars in your factory (a huge machine or machine shop) will go through the same machines. So you need machines to cater to individual customer selections or unique customer selections. In our project not all models have animations but the function of the component was designed to handle those rare cases. This is why we have a condition to check if the model has animations in the first place. Again we might base faced with edge cases as we move on to the topic of the number of clips available in the animation array so we iterate through it. This checks the array for one clip or many of them. It doesn’t discriminate. The rest is just a matter of mixer object and properties - nothing particularly complex. Lastly, the variable of type let was used because it needs to be available to other stuff outside of its scope and the editor doesn’t allow a const type and variable name to stand on its own anyway. Oops, not finished quite yet. Note the use of an optional chaining operator after mixer in the dot notation. This is again a part of the preemptive design of code.